// pronti-frontend-main/front/src/app/modulos/registro-vacacion/form/form-registro-vacacion.enum.ts
export enum FormRegistroVacacionEnum {
  // nombre = 'nombre',
  fechaDesde = 'fechaDesde',
  fechaHasta = 'fechaHasta',
  diasTomados = 'diasTomados',
  valorTomado = 'valorTomado',
  estaPagado = 'estaPagado',
  nombreApellidoResponsable = 'nombreApellidoResponsable',
  fechaPago = 'fechaPago',
  comprobantePago = 'comprobantePago',
  valorPagado = 'valorPagado',
}